+++ include/OrganEngine/Ringbuffer.h
@@ -8,19 +8,20 @@
 #endif
 
 #include <vector>
+#include <cstdint>
 
 // Hermite polynomial interpolation. More points - better response at high frequencies
 __attribute__((always_inline)) inline
-float getSampleHermite4p3o(float x, float *y)
+int16_t getSampleHermite4p3o(float x, int16_t *y)
 {
     static float c0, c1, c2, c3;
 
     // 4-point, 3rd-order Hermite (x-form)
-    c0 = y[1];
-    c1 = (1.0/2.0)*(y[2]-y[0]);
-    c2 = (y[0] - (5.0/2.0)*y[1]) + (2.0*y[2] - (1.0/2.0)*y[3]);
-    c3 = (1.0/2.0)*(y[3]-y[0]) + (3.0/2.0)*(y[1]-y[2]);
-    return ((c3*x+c2)*x+c1)*x+c0;
+    c0 = (float)y[1];
+    c1 = (1.0/2.0)*((float)y[2]-(float)y[0]);
+    c2 = ((float)y[0] - (5.0/2.0)*(float)y[1]) + (2.0*(float)y[2] - (1.0/2.0)*(float)y[3]);
+    c3 = (1.0/2.0)*((float)y[3]-(float)y[0]) + (3.0/2.0)*((float)y[1]-(float)y[2]);
+    return (int16_t) (((c3*x+c2)*x+c1)*x+c0);
 }
 
 class RingBuffer
@@ -29,8 +30,8 @@ public:
 	RingBuffer();
 
     void write(float sample);
-    void write_margined(float sample);
-    float readWithDelay(int delay);
+    void write_margined(int16_t sample);
+    int16_t readWithDelay(int delay);
 	void resize(int size);
 
 public:
@@ -50,7 +51,7 @@ public:
     }
 
 private:
-	std::vector<float>  buffer;
+	std::vector<int16_t>  buffer;
 	int                 writeIndex;
 	int                 size;
